TIP-860: Move Family related api tests
