+++ tests/back/Pim/Structure/EndToEnd/Family/ExternalApi/GetFamilyEndToEnd.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Family;
+namespace AkeneoTest\Pim\Structure\EndToEnd\Family\ExternalApi;
 
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class GetFamilyIntegration extends ApiTestCase
+class GetFamilyEndToEnd extends ApiTestCase
 {
     public function testGetAFamily()
     {
