+++ tests/back/Pim/Structure/EndToEnd/Family/ExternalApi/PartialUpdateListFamilyEndToEnd.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Family;
+namespace AkeneoTest\Pim\Structure\EndToEnd\Family\ExternalApi;
 
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\Stream\StreamResourceResponse;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class PartialUpdateListFamilyIntegration extends ApiTestCase
+class PartialUpdateListFamilyEndToEnd extends ApiTestCase
 {
     public function testCreateAndUpdateAListOfFamilies()
     {
